Eliminate function pointers in hpack_enc().
Using a conditional branch to set a function pointer will prevent the
compiler to inline the function and worse it would jump based on a
register value killing the pipline.

Here is a short example to demonstrate the variants:
https://godbolt.org/z/radfZg

Remove `add_nothing` and add a lambda when necessary to make sure
everything is inlined.