docs: enumerate status codes in statuscodes.md
statuscodes.md is one of the top search results for "grpc status codes",
right behind codes.proto. Since this is a landing page for many people*,
and since the title and description of this page purports to be the
generic status codes documentation page, it seems like a good idea to define
terms before using them.

* Many people who are new to grpc/protobuf, for example, will be more
accustomed to and choose markdown files over .proto files.