Deflake QPS test.
In https://github.com/grpc/grpc/pull/20328, we started shutting down the
servers and the clients at the time to make the benchmark resilient.

Since servers and clients are shutdown in parallel, there is always a
chance that they observe cancelled RPCs. This resulted in QPS benchmarks
failing.  In this commit, we also accept cancelled RPCs in addition to
OK RPCs.

Fixes #20509